match: reconstruct PATH ValidAttack facing domain

+++ src/SOURCE/PATH.cpp
@@ -218,14 +218,16 @@ i32 army::ValidAttack(
         if (direction == COMBAT_DIRECTION_WIDE_WEST) {
             *attackHex = GetAdjacentCellIndex(
                 sourceHex,
-                m_facing < ARMY_FACING_RIGHT ? COMBAT_DIRECTION_NORTHWEST
-                                                               : COMBAT_DIRECTION_NORTHEAST
+                static_cast<u32>(m_facing) < static_cast<u32>(ARMY_FACING_RIGHT)
+                    ? COMBAT_DIRECTION_NORTHWEST
+                    : COMBAT_DIRECTION_NORTHEAST
             );
         } else if (direction == COMBAT_DIRECTION_WIDE_EAST) {
             *attackHex = GetAdjacentCellIndex(
                 sourceHex,
-                m_facing < ARMY_FACING_RIGHT ? COMBAT_DIRECTION_SOUTHWEST
-                                                               : COMBAT_DIRECTION_SOUTHEAST
+                static_cast<u32>(m_facing) < static_cast<u32>(ARMY_FACING_RIGHT)
+                    ? COMBAT_DIRECTION_SOUTHWEST
+                    : COMBAT_DIRECTION_SOUTHEAST
             );
         } else {
             switch (m_facing) {
